FILE: renames create-tree to create-root
